Proceeding towards dockerization and TDD
Major changes in this commit. Firstly, feature enhancement will be done according to Test Driven Development (TDD). For this, the C testing framework called 'check' is being used. Secondly, because this method and the framework requires several changes, it is best to do it within a dockerized enviroment. Hence, VSCode's excellent feature called 'devcontainer' is being used.