Interface bug fixes and completed v1.0 of this small project
The input is expected to be of int type but this was not being checked. This has now been fixed and is also included in testing.
The project diverted quite a lot from what I had expected it to be. But for now, this will be considered as a release. I might come back to this later and improve it or open a new repo to implement what was wanted.